SQUAD-9-313 Adding validation and handling exceptions for 'delete client' request

+++ src/controllers/ClientsController.js
@@ -1,5 +1,6 @@
 const ClientModel = require('../models/ClientModel');
 const validator = require('../helpers/validator');
+const { NotFoundError } = require('../errors');
 
 class ClientsController {
 	// GET - Returns a list of clients
@@ -24,10 +25,19 @@ class ClientsController {
 	// DELETE - Delete a client
 	static async deleteOne(req) {
 		const { clientId } = req.params;
+		let checkClientExists = await ClientModel.getOne(clientId);
 
-		await ClientModel.deleteById(clientId);
+		if (checkClientExists == undefined)
+			throw new NotFoundError("Error occured while retrieving details of" +
+															" client Id: " + clientId);
 
-		return { message: 'success' };
+		let deleteByIdResult = await ClientModel.deleteById(clientId);
+
+		if (deleteByIdResult)
+			throw new Error("Error occured while deleting the client Id: " +
+											clientId);
+
+		return { message: "Client Id: " + clientId + "deleted successfully" };
 	}
 
 	// PUT - Update a client
